backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
redis==5.0.1
celery==5.3.4
markdown==3.5.1
bleach==6.1.0
aiofiles==23.2.1
python-slugify==8.0.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
asyncpg==0.30.0
# New dependencies for v2.0
minio==7.2.0
PyYAML==6.0.1
python-frontmatter==1.0.1
whoosh==2.7.4